Fix refactoring merge
