+++ app/src/main/java/de/fhdw/ergoholics/brainphaser/activities/fileimport/ImportChallengeActivity.java
@@ -11,23 +11,16 @@ import android.view.View;
 import de.fhdw.ergoholics.brainphaser.BrainPhaserApplication;
 import de.fhdw.ergoholics.brainphaser.R;
 import de.fhdw.ergoholics.brainphaser.activities.main.MainActivity;
-import de.fhdw.ergoholics.brainphaser.fileimport.FileImport;
-import de.fhdw.ergoholics.brainphaser.fileimport.exceptions.ElementAmountException;
-import de.fhdw.ergoholics.brainphaser.fileimport.exceptions.FileFormatException;
-import de.fhdw.ergoholics.brainphaser.fileimport.exceptions.InvalidAttributeException;
-import de.fhdw.ergoholics.brainphaser.fileimport.exceptions.UnexpectedElementException;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.FileImport;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.ElementAmountException;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.FileFormatException;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.InvalidAttributeException;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.UnexpectedElementException;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.InputStream;
-
-import de.fhdw.ergoholics.brainphaser.BrainPhaserApplication;
-import de.fhdw.ergoholics.brainphaser.R;
-import de.fhdw.ergoholics.brainphaser.activities.main.MainActivity;
-import de.fhdw.ergoholics.brainphaser.logic.fileimport.FileImport;
-import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.FileFormatException;
-import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.UnexpectedElementException;
 
 /**
  * Created by Daniel Hoogen on 16/02/2016.
