+++ app/src/main/java/de/fhdw/ergoholics/brainphaser/logic/fileimport/FileImport.java
@@ -1,24 +1,22 @@
-package de.fhdw.ergoholics.brainphaser.fileimport;
-
-import org.w3c.dom.Node;
-
-import java.io.File;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
+package de.fhdw.ergoholics.brainphaser.logic.fileimport;
 
 import de.fhdw.ergoholics.brainphaser.BrainPhaserApplication;
-import de.fhdw.ergoholics.brainphaser.activities.BrainPhaserActivity;
-import de.fhdw.ergoholics.brainphaser.fileimport.bpc.BPCObjects;
-import de.fhdw.ergoholics.brainphaser.fileimport.bpc.BPCRead;
-import de.fhdw.ergoholics.brainphaser.fileimport.bpc.BPCWrite;
-import de.fhdw.ergoholics.brainphaser.fileimport.exceptions.ElementAmountException;
-import de.fhdw.ergoholics.brainphaser.fileimport.exceptions.FileFormatException;
-import de.fhdw.ergoholics.brainphaser.fileimport.exceptions.InvalidAttributeException;
-import de.fhdw.ergoholics.brainphaser.fileimport.exceptions.UnexpectedElementException;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.bpc.BPCObjects;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.bpc.BPCRead;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.bpc.BPCWrite;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.ElementAmountException;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.FileFormatException;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.InvalidAttributeException;
+import de.fhdw.ergoholics.brainphaser.logic.fileimport.exceptions.UnexpectedElementException;
 import de.fhdw.ergoholics.brainphaser.model.Answer;
 import de.fhdw.ergoholics.brainphaser.model.Category;
 import de.fhdw.ergoholics.brainphaser.model.Challenge;
+
+import org.w3c.dom.Node;
+
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Created by Daniel Hoogen on 19/02/2016.
@@ -56,7 +54,7 @@ public class FileImport {
         while (childCategories!=null) {
             if (childCategories.getNodeType()==Node.ELEMENT_NODE) {
                 nextChallengeId = BPCObjects.readCategory(childCategories, i,
-                        nextChallengeId, categoryList, challengeList, answerList);
+                    nextChallengeId, categoryList, challengeList, answerList);
                 i++;
             }
